// src/types.ts
export interface Subject {
  id: string;
  name: string;
  color: string;
  isArchived?: boolean;
}
// ...
export interface Exam {
  id: string;
  subjectIds: string[]; // Changed from subjectId to subjectIds
  title: string;
  date: string; // ISO Date YYYY-MM-DD
  topics: string;
  createdAt: number;
}